core/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Max Halford <maxhalford25@gmail.com>"]
readme = "README.md"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
pytest-xdist = "^3.1.0"
pre-commit = "^3.0.0"
mypy = "^0.991"
pydantic = "^1.10.2"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
sqlmodel = "^0.0.8"
uvicorn = "^0.20.0"
kafka-python = "^2.0.2"
dill = "^0.3.6"
psycopg = {extras = ["binary"], version = "^3.1.4"}
psycopg2 = "^2.9.5"
celery = {extras = ["redis"], version = "^5.2.7"}
river = "^0.15.0"
websockets = "^10.4"

[tool.poetry.group.dev.dependencies]
httpx = "^0.23.3"

[tool.ruff]
line-length = 100
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = [
    "--doctest-modules",
    "--doctest-glob=README.md",
    "--verbose",
    "-ra",
    "--color=yes",
]
doctest_optionflags = "NORMALIZE_WHITESPACE NUMBER ELLIPSIS"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
